ecommerce-backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
